# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "The single outbound-messaging boundary: in-app, push, email and WhatsApp, plus OTP and support tickets.\\n\\nMost traffic arrives over gRPC from other services. These HTTP endpoints serve a user's own inbox, the OTP flow, and the WhatsApp webhook.",
        "title": "Karlo Notification API",
        "termsOfService": "https://karlo.co.id/terms",
        "contact": {
            "name": "Karlo Engineering",
            "email": "engineering@karlo.co.id"
        },
        "version": "1.0"
    },
    "host": "localhost:5004",
    "basePath": "/api/v1",
    "paths": {
        "/notifications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Notifications"
                ],
                "summary": "List notifications",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_karlo_notification-service_internal_platform_response.Meta"
                        }
                    }
                }
            }
        },
        "/notifications/read": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Notifications"
                ],
                "summary": "Mark notifications read",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/notifications/unread": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Notifications"
                ],
                "summary": "Count unread notifications",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/otp/send": {
            "post": {
                "tags": [
                    "OTP"
                ],
                "summary": "Send an OTP",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/otp/verify": {
            "post": {
                "tags": [
                    "OTP"
                ],
                "summary": "Verify an OTP",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/webhooks/whatsapp": {
            "get": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Verify the WhatsApp webhook",
                "responses": {}
            },
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Receive WhatsApp events",
                "responses": {}
            }
        }
    },
    "definitions": {
        "github_com_karlo_notification-service_internal_platform_response.Meta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                },
                "totalRows": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "RS256 access token issued by the authentication service, as \"Bearer \u003ctoken\u003e\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}